+++ tests/system/data_simulator.py
@@ -9,8 +9,8 @@ import numpy as np
 from dummydata.model2 import Model2
 from dummydata.model3 import Model3
 
-from esmvaltool.data_finder import get_input_filename
 from esmvaltool._config import read_config_user_file
+from esmvaltool._data_finder import get_input_filename
 from esmvaltool._namelist import read_namelist_file
